nacos学习第2节

+++ payment-provider9002/src/main/java/com/jfa/controller/PaymentController.java
@@ -9,11 +9,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class PaymentController {
 
+
+
     @Value("${server.port}")
     private String serverPort;
 
     @GetMapping(value = "/payment/nacos/{id}")
-    public String getPayment(@PathVariable("id") Integer id){
+    public String getPayment(@PathVariable("id") Long id){
         return "nacos registry,serverPort:"+serverPort+"\t id"+id;
     }
 }
